Fix clippy warnings
- Fix approx_constant error (use PI constant)
- Remove broken benches/search.rs
- Auto-fix redundant closures, borrows, unnecessary casts
- Fix manual_strip to use strip_prefix
- Fix vec_init_then_push to use vec![]
- Fix ptr_arg (&PathBuf -> &Path)
- Fix unnecessary_filter_map
- Fix manual_clamp
- Fix redundant_pattern_matching
- Fix field_reassign_with_default
- Fix suspicious_open_options
- Add type alias for complex type
- Add #[allow(dead_code)] for incomplete features

+++ src/system/tray.rs
@@ -10,7 +10,7 @@ pub fn create_tray_icon() -> Result<TrayIcon> {
         .into_rgba8();
     let (width, height) = image.dimensions();
     let rgba = image.into_raw();
-    let icon = Icon::from_rgba(rgba, width as u32, height as u32)
+    let icon = Icon::from_rgba(rgba, width, height)
         .map_err(|e| FlashError::config("tray_icon", e.to_string()))?;
 
     let tray_menu = Menu::new();
